git-svn-id: svn://toshia.dip.jp/mikutter/trunk@87 03aab468-d3d2-4883-8b12-f661bbf03fa8

+++ mikutter.rb
@@ -159,7 +159,7 @@ argument_parser()
 begin
   $stderr = File.open(errfile, 'w') if not $debug
   boot()
-  File.delete(errfile)
+  File.delete(errfile) if File.exist?(errfile)
 ensure
   # $stderr.close if errlog.closed?
 end
